added popups folder, moved Popup and PopupWrapper into it, updated indexes
