--- src/modules/common/PopupWrapper/PopupWrapper.scss
@@ -1,7 +1,0 @@
-.popup-wrapper {
-  width: 100%;
-
-  &__content {
-    user-select: none;
-  }
-}
