+++ src/modules/common/popups/PopupWrapper/PopupWrapper.scss
@@ -1,0 +1,15 @@
+.popup-wrapper {
+  position: relative;
+
+  width: 100%;
+
+  &__content {
+    user-select: none;
+
+    & > * {
+      width: 100%;
+
+      box-sizing: border-box;
+    }
+  }
+}
