+++ src/modules/common/popups/index.js
@@ -1,0 +1,4 @@
+import Popup from './Popup';
+import PopupWrapper from './PopupWrapper';
+
+export { Popup, PopupWrapper };
